rename my composant , and fix position side bar
