--- src/components/navbar/Navbar.css
@@ -1,5 +1,0 @@
-.navbar {
-    background-color: rgb(130, 174, 227);
-    border-right-style: solid;
-    height: 100%;
-}
