--- src/components/navbar/Navbar.tsx
@@ -1,11 +1,0 @@
-"use client";
-import React from "react";
-import "./Navbar.css";
-
-type typeNavbar = {
-  children: React.ReactNode;
-};
-
-export const Navbar = ({ children }: typeNavbar) => {
-  return <nav className="navbar">{children}</nav>;
-};
