+++ src/components/sideBar/SideBar.css
@@ -1,0 +1,18 @@
+.sidebar {
+    background-color: rgb(130, 174, 227);
+    border-right-style: solid;
+    height: 100%;
+    width: 25%;
+}
+.sidebar.p20 {
+    padding:20px;
+}
+@media (min-width: 340px) and (max-width: 950px) {
+    .sidebar {
+        position: fixed;
+        left: 0;
+        bottom: 0;
+        width:100%;
+        height:20%;
+    }
+}
