+++ src/container/sideBar/SideBar.css
@@ -1,0 +1,52 @@
+.git-logo {
+    display: flex;
+    justify-content: center;
+    padding-top: 20px;
+    font-size: 3em;
+}
+
+.liste{
+    display: flex;
+    /* justify-content: center; */
+    align-items: center;
+    padding-top: 20px;
+    padding-bottom: 20px;
+}
+
+.lists {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    padding-top: 3em;
+}
+
+.nav-logo {
+    font-size:xx-large;
+    padding-left: 20px;
+    padding-right:20px;
+}
+
+.title {
+    font-size: x-large;
+    font-family: 'Inter', 'Roboto', monospace;
+}
+
+.activeLink {
+    border-radius: 10px;
+    background-color: rgb(192, 178, 234);
+}
+
+@media (min-width: 340px) and (max-width: 950px) {
+    .git-logo {
+       display: none;
+    }
+    .lists{
+        display: flex;
+        flex-direction: row;
+        justify-content: space-around;
+    }
+    .title{
+        display:none;
+    }
+}
+
